Pitch cam control by AUX Channel
I'm not a programmer in ASM but do you think it's possible to link de AUX channel to control the pitch of the servopitch offset.  So we can fix in front of the quad when de AUX channel is at 50%, and look upward or downward when tuning it to 0 or 100%.  Of course, the Pitch/Roll Servo should be able to look so high or low?